requirements.txt
numpy==1.18.1
scikit-image==0.16.2
Shapely==1.7.0
tensorflow-gpu==2.3.0
tqdm==4.48.2
